# docs/swagger.yaml
basePath: /api
definitions:
  activity.SActivityLogConfigReq:
    properties:
      display_name:
        type: string
      fields:
        items:
          $ref: '#/definitions/activity.SActivityLogFieldReq'
        type: array
      is_enabled:
        type: boolean
      log_create:
        type: boolean
      log_delete:
        type: boolean
      log_update:
        type: boolean
      primary_key_field:
        type: string
      source_name:
        type: string
      table_name:
        type: string
      target_table:
        type: string
    required:
    - fields
    - table_name
    - target_table
    type: object
  activity.SActivityLogFieldReq:
    properties:
      display_name:
        type: string
      field_name:
        type: string
      field_type:
        type: string
      is_sensitive:
        type: boolean
      is_tracked:
        type: boolean
      sort_order:
        type: integer
    required:
    - field_name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for DAPEN management of employee retirement funds
  title: DAPEN System API
  version: "1.0"
paths:
  /admin/activity-logs:
    get:
      description: Fetch a paginated list of activity logs
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Activity Logs
      tags:
      - Activity Log
  /admin/activity-logs/configs:
    get:
      description: Fetch all activity log configurations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Activity Log Configs
      tags:
      - Activity Log
    post:
      consumes:
      - application/json
      description: Create or update an activity log configuration
      parameters:
      - description: Configuration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/activity.SActivityLogConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Save Activity Log Config
      tags:
      - Activity Log
  /admin/activity-logs/configs/{table_name}:
    get:
      description: Fetch activity log configuration for a specific table
      parameters:
      - description: Table Name
        in: path
        name: table_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Log Config by Table Name
      tags:
      - Activity Log
  /admin/database/tables:
    get:
      description: Fetch a list of all tables in the database for configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get All Database Tables
      tags:
      - Database
  /admin/database/tables/{table_name}/columns:
    get:
      description: Fetch a list of columns for a specific database table
      parameters:
      - description: Table Name
        in: path
        name: table_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Columns for Table
      tags:
      - Database
...
  /dashboard/pensiunan-without-files:
    get:
      description: Fetch list of retired customers without uploaded files
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Pensiunan Without Files
      tags:
      - Dashboard
  /dashboard/stats:
    get:
      description: Fetch various statistics for the dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Dashboard Statistics
      tags:
      - Dashboard
...
  /menus/sidebar:
    get:
      description: Fetch dynamic sidebar menu based on user permissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Sidebar Menu
      tags:
      - Dashboard
...
swagger: "2.0"